Consolidate JS files under single `js/` directory.
There's little point to having lots of directories with just 1-2 files in them each. For bigger groups (in fbsdkshare) we can have subdirectories (e.g. `js/models`).
