+++ react-native-fbsdkcore/js/FBSDKAccessToken.ios.js
