+++ react-native-fbsdkcore/js/FBSDKAppEvents.ios.js
