+++ react-native-fbsdkcore/js/FBSDKGraphRequestManager.ios.js
