+++ react-native-fbsdklogin/js/FBSDKLoginManager.ios.js
