+++ react-native-fbsdkshare/FBSDKShare.ios.js
@@ -23,31 +23,31 @@
 'use strict';
 
 // Share UI components
-exports.FBSDKShareButton = require('./js-ui/FBSDKShareButton.ios.js');
-exports.FBSDKSendButton = require('./js-ui/FBSDKSendButton.ios.js');
-exports.FBSDKLikeControl = require('./js-ui/FBSDKLikeControl.ios.js');
+exports.FBSDKShareButton = require('./js/ui/FBSDKShareButton.ios.js');
+exports.FBSDKSendButton = require('./js/ui/FBSDKSendButton.ios.js');
+exports.FBSDKLikeControl = require('./js/ui/FBSDKLikeControl.ios.js');
 
 // Share dialogs
-exports.FBSDKShareDialog = require('./js-modules/FBSDKShareDialog.ios.js');
-exports.FBSDKMessageDialog = require('./js-modules/FBSDKMessageDialog.ios.js');
-exports.FBSDKAppGroupAddDialog = require('./js-modules/FBSDKAppGroupAddDialog.ios.js');
-exports.FBSDKAppGroupJoinDialog = require('./js-modules/FBSDKAppGroupJoinDialog.ios.js');
-exports.FBSDKAppInviteDialog = require('./js-modules/FBSDKAppInviteDialog.ios.js');
-exports.FBSDKGameRequestDialog = require('./js-modules/FBSDKGameRequestDialog.ios.js');
+exports.FBSDKShareDialog = require('./js/FBSDKShareDialog.ios.js');
+exports.FBSDKMessageDialog = require('./js/FBSDKMessageDialog.ios.js');
+exports.FBSDKAppGroupAddDialog = require('./js/FBSDKAppGroupAddDialog.ios.js');
+exports.FBSDKAppGroupJoinDialog = require('./js/FBSDKAppGroupJoinDialog.ios.js');
+exports.FBSDKAppInviteDialog = require('./js/FBSDKAppInviteDialog.ios.js');
+exports.FBSDKGameRequestDialog = require('./js/FBSDKGameRequestDialog.ios.js');
+
+// API
+exports.FBSDKShareAPI = require('./js/FBSDKShareAPI.ios.js');
 
 // Share content models
-exports.FBSDKSharingContent = require('./js-models/FBSDKSharingContent.ios.js');
-exports.FBSDKSharePhoto = require('./js-models/FBSDKSharePhoto.ios.js');
-exports.FBSDKSharePhotoContent = require('./js-models/FBSDKSharePhotoContent.ios.js');
-exports.FBSDKShareVideo = require('./js-models/FBSDKShareVideo.ios.js');
-exports.FBSDKShareVideoContent = require('./js-models/FBSDKShareVideoContent.ios.js');
-exports.FBSDKShareLinkContent = require('./js-models/FBSDKShareLinkContent.ios.js');
-exports.FBSDKShareOpenGraphContent = require('./js-models/FBSDKShareOpenGraphContent.ios.js');
-exports.FBSDKShareOpenGraphAction = require('./js-models/FBSDKShareOpenGraphAction.ios.js');
-exports.FBSDKShareOpenGraphObject = require('./js-models/FBSDKShareOpenGraphObject.ios.js');
-exports.FBSDKAppGroupAddContent = require('./js-models/FBSDKAppGroupAddContent.ios.js');
-exports.FBSDKAppInviteContent = require('./js-models/FBSDKAppInviteContent.ios.js');
-exports.FBSDKGameRequestContent = require('./js-models/FBSDKGameRequestContent.ios.js');
-
-// Misc.
-exports.FBSDKShareAPI = require('./js-modules/FBSDKShareAPI.ios.js');
+exports.FBSDKSharingContent = require('./js/models/FBSDKSharingContent.ios.js');
+exports.FBSDKSharePhoto = require('./js/models/FBSDKSharePhoto.ios.js');
+exports.FBSDKSharePhotoContent = require('./js/models/FBSDKSharePhotoContent.ios.js');
+exports.FBSDKShareVideo = require('./js/models/FBSDKShareVideo.ios.js');
+exports.FBSDKShareVideoContent = require('./js/models/FBSDKShareVideoContent.ios.js');
+exports.FBSDKShareLinkContent = require('./js/models/FBSDKShareLinkContent.ios.js');
+exports.FBSDKShareOpenGraphContent = require('./js/models/FBSDKShareOpenGraphContent.ios.js');
+exports.FBSDKShareOpenGraphAction = require('./js/models/FBSDKShareOpenGraphAction.ios.js');
+exports.FBSDKShareOpenGraphObject = require('./js/models/FBSDKShareOpenGraphObject.ios.js');
+exports.FBSDKAppGroupAddContent = require('./js/models/FBSDKAppGroupAddContent.ios.js');
+exports.FBSDKAppInviteContent = require('./js/models/FBSDKAppInviteContent.ios.js');
+exports.FBSDKGameRequestContent = require('./js/models/FBSDKGameRequestContent.ios.js');
