+++ react-native-fbsdkshare/js/FBSDKAppGroupJoinDialog.ios.js
