+++ react-native-fbsdkshare/js/FBSDKShareTypes.ios.js
