+++ react-native-fbsdkshare/js/models/FBSDKAppGroupAddContent.ios.js
