+++ react-native-fbsdkshare/js/models/FBSDKAppInviteContent.ios.js
