+++ react-native-fbsdkshare/js/models/FBSDKGameRequestContent.ios.js
