+++ react-native-fbsdkshare/js/models/FBSDKSharePhoto.ios.js
