+++ react-native-fbsdkshare/js/models/FBSDKSharePhotoContent.ios.js
