+++ react-native-fbsdkshare/js/models/FBSDKSharingContent.ios.js
