+++ react-native-fbsdkshare/js/ui/FBSDKLikeControl.ios.js
