+++ react-native-fbsdkshare/js/ui/FBSDKShareButton.ios.js
@@ -22,15 +22,48 @@
 
 'use strict';
 
-var FBSDKAppGroupAddDialogInterface = require('react-native').NativeModules.FBSDKAppGroupAddDialog;
+var React = require('react-native');
+var {
+  requireNativeComponent,
+  StyleSheet,
+} = React;
 
-import type * as FBSDKAppGroupAddContent from '../js-models/FBSDKAppGroupAddContent.ios.js';
-import type {
-  FBSDKDialogCallback,
-} from './FBSDKShareTypes.ios.js';
+var FBSDKSharingContent = require('../models/FBSDKSharingContent.ios.js');
 
-module.exports = {
-  show(content: FBSDKAppGroupAddContent, callback: FBSDKDialogCallback) {
-    FBSDKAppGroupAddDialogInterface.show(content, callback);
-  },
+/**
+ * A button to share content.
+ */
+class FBSDKShareButton extends React.Component {
+  render() {
+    return (
+      <RCTFBSDKShareButton
+        {...this.props}
+        style={this.props.style || styles.fbsdkShareButton}
+      />
+    );
+  }
+}
+
+FBSDKShareButton.propTypes = {
+  /**
+   * The content to be shared.
+   */
+  shareContent: React.PropTypes.instanceOf(FBSDKSharingContent),
 };
+
+var styles = StyleSheet.create({
+  /**
+   * Default styling for the share button
+   */
+  fbsdkShareButton: {
+    width: 100,
+    height: 25,
+  },
+});
+
+var RCTFBSDKShareButton = requireNativeComponent(
+  'RCTFBSDKShareButton',
+  FBSDKShareButton
+);
+
+module.exports = FBSDKShareButton;
